# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Class Scheduling API
  version: 1.0.0
  description: >
    REST API for managing 1-hour class slots for a single teacher. Supports students,
    slots (book/cancel/free/replace/reschedule), blocks (unavailability), audit events,
    and an optional waitlist.
servers:
  - url: /api/v1
...
paths:
...
  /waitlist:
    post:
      tags: [Waitlist]
      operationId: addToWaitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitlistCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    get:
      tags: [Waitlist]
      operationId: listWaitlist
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistPageResponse'
        '400':
          $ref: '#/components/responses/Problem'
  /waitlist/{waitlistId}:
    delete:
      tags: [Waitlist]
      operationId: removeFromWaitlist
      parameters:
        - $ref: '#/components/parameters/WaitlistId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/Problem'
components:
  parameters:
...
    WaitlistId:
      name: waitlistId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Spring-style sort, e.g. startAt,asc
      schema:
        type: string
        default: "startAt,asc"
  responses:
    Problem:
      description: Problem Details (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          example: "https://example.com/problems/validation-error"
        title:
          type: string
          example: "Validation error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "startAt must be in the future"
        instance:
          type: string
          example: "/api/v1/slots"
        errors:
          type: array
          description: Optional list of field errors
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: "startAt"
        message:
          type: string
          example: "must not be null"
...
    StudentBrief:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
...
    WaitlistCreateRequest:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string
          format: uuid
        preferredDays:
          type: array
          items:
            $ref: '#/components/schemas/DayOfWeek'
          nullable: true
        preferredTimeRanges:
          type: array
          items:
            $ref: '#/components/schemas/TimeRange'
          nullable: true
        notes:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          type: integer
          default: 0
    TimeRange:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          example: "16:00"
        to:
          type: string
          example: "20:00"
    WaitlistResponse:
      type: object
      required: [id, student, active, createdAt, priority]
      properties:
        id:
          type: string
          format: uuid
        student:
          $ref: '#/components/schemas/StudentBrief'
        preferredDays:
          type: array
          items:
            $ref: '#/components/schemas/DayOfWeek'
          nullable: true
        preferredTimeRanges:
          type: array
          items:
            $ref: '#/components/schemas/TimeRange'
          nullable: true
        notes:
          type: string
          nullable: true
        priority:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WaitlistPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WaitlistResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer